nvim (fix): Diagnostics overriding hover window
I like to have a diagnostic popup window automatically show
whenever I hover over a symbol. It is often the case that
I also want to see the documentation of that symbol using
vim.lsp.buf.hover(). The problem is that the diagnostic
popup will "override" the hover window making it unreadable.

This fix was found on Reddit and I'm not sure if it will
cause any problems in the long run, but for now, it gets the
job done.